requirements.txt
Django==5.2.4
django-environ==0.12.0
django-ckeditor-5==0.2.18
Pillow==11.3.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
gunicorn==23.0.0
boto3==1.34.0
django-storages==1.14.2
django-redis==6.0.0
redis==5.2.1
hiredis==3.3.0